Moved syncer module to a common utils package.
