+++ src/aturtle/utils/__init__.py
@@ -1,0 +1,8 @@
+# ----------------------------------------------------------------------------
+# Python A-Turtle
+# ----------------------------------------------------------------------------
+# Copyright (c) Tiago Montes.
+# See LICENSE for details.
+# ----------------------------------------------------------------------------
+
+from . import syncer
